+++ protected/humhub/modules/activity/views/layouts/mail_plaintext.php
@@ -21,5 +21,5 @@ use humhub\modules\space\models\Space;
 <?php endif; ?>
 
 <?php if ($url != '') : ?>
-    <?= Yii::t('ActivityModule.base', 'See online:') ?> <?= urldecode($url) ?>
+    <?= Yii::t('ActivityModule.base', 'See online:') ?> <?= urldecode((string) $url) ?>
 <?php endif; ?>