+++ protected/humhub/modules/content/Events.php
@@ -86,15 +86,17 @@ class Events extends BaseObject
         $integrityController->showTestHeadline('Content Objects (' . Content::find()->count() . ' entries)');
         foreach (Content::find()->each() as $content) {
             /* @var Content $content */
-            if ($content->createdBy == null) {
-                if ($integrityController->showFix('Deleting content id ' . $content->id . ' of type ' . $content->object_model . ' without valid user!')) {
-                    $content->getPolymorphicRelation()->hardDelete();
-                }
+            if (
+                !$content->getCreatedBy()->exists()
+                && $integrityController->showFix('Deleting content id ' . $content->id . ' of type ' . $content->object_model . ' without valid user!')
+            ) {
+                $content->getPolymorphicRelation()->hardDelete();
             }
-            if ($content->getPolymorphicRelation() === null) {
-                if ($integrityController->showFix('Deleting content id ' . $content->id . ' of type ' . $content->object_model . ' without valid content object!')) {
-                    $content->getPolymorphicRelation()->hardDelete();
-                }
+            if (
+                $content->getPolymorphicRelation() === null
+                && $integrityController->showFix('Deleting content id ' . $content->id . ' of type ' . $content->object_model . ' without valid content object!')
+            ) {
+                $content->hardDeleteInternal();
             }
         }
     }