+++ protected/humhub/modules/notification/views/layouts/mail_plaintext.php
@@ -1,1 +1,1 @@
-<?= $content;
+<?php echo $content;