+++ CHANGELOG.md
@@ -15,6 +15,9 @@ HumHub Changelog
 - Enh #8083: Store module version in DB
 - Enh #5141: Added CLI command to delete disabled or soft deleted users
 - Enh #8100: Move `static` and `themes` into `protected/humhub`
+- Enh #8128: Exact search for user id in admin user search
+- Enh #8129: Limit activity content length
+- Enh #8139: Update rector rules
 
 1.18.3 (TBD)
 ------------